backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "EcoTrace backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
]
